iFix the bug of N penalty.

+++ src/ssw_cpp.cpp
@@ -38,12 +38,12 @@ void BuildSwScoreMatrix(const uint8_t& match_score,
       matrix[id] = ((i == j) ? match_score : static_cast<int8_t>(-mismatch_penalty));
       ++id;
     }
-    matrix[id] = -mismatch_penalty; // For N
+    matrix[id] = static_cast<int8_t>(-mismatch_penalty); // For N
     ++id;
   }
 
   for (int i = 0; i < 5; ++i)
-    matrix[id++] = -mismatch_penalty; // For N
+    matrix[id++] = static_cast<int8_t>(-mismatch_penalty); // For N
 
 }
